Melhorias na pagina de detalhamento do anuncio. Preview das imagens carregadas na pagina de criação de anuncio.

+++ MDN/MDN/Models/Entities/T001_PRODUTO.cs
@@ -13,7 +13,7 @@ namespace MDN.Models.Entities
         public int T001_ID_PRODUTO { get; set; }
         public string T001_TITULO { get; set; }
         public string T001_DESCRICAO { get; set; }
-        public decimal T001_PRECO { get; set; }
+        public int T001_PRECO { get; set; }
         public DateTime T001_DT_CRIACAO { get; set; }
         public bool T001_ATIVO { get; set; }
         public int T003_ID_UF { get; set; }
